Change semId type to uint8_t.

+++ Userland/SampleCodeModule/include/my_sync.h
@@ -5,12 +5,12 @@
 
 void sem_dump(void);
 
-int sem_open(uint16_t semId, uint64_t value);
+int sem_open(uint8_t semId, uint64_t value);
 
-int sem_wait(uint16_t semId);
+int sem_wait(uint8_t semId);
 
-int sem_post(uint16_t semId);
+int sem_post(uint8_t semId);
 
-int sem_close(uint16_t semId);
+int sem_close(uint8_t semId);
 
 #endif
